+++ isort/main.py
@@ -588,7 +588,16 @@ def _build_arg_parser() -> argparse.ArgumentParser:
             " aliases to signify intent and change behaviour."
         ),
     )
-
+    parser.add_argument(
+        "--float-to-top",
+        dest="float_to_top",
+        action="store_true",
+        help="Causes all non indented imports to float to the top of the file having its imports "
+        "sorted. *NOTE*: This is a **beta** feature. It currently doesn't work with cimports and "
+        "is gauranteed to run much slower and use much more memory than the default. Still it "
+        "can be a great shortcut for collecting imports every once in a while when you put them "
+        "in the middle of a file.",
+    )
     # deprecated options
     parser.add_argument(
         "--recursive",