+++ CHANGELOG.md
@@ -6,6 +6,7 @@ NOTE: isort follows the [semver](https://semver.org/) versioning standard.
 ### 5.1.0 July TBD, 2020
   - isort now throws an exception if an invalid settings path is given (issue #1174).
   - Implemented support for automatic redundant alias removal (issue #1281).
+  - Implemented experimental support for floating all imports to the top of a file (issue #1228)
   - Fixed #1178: support for semicolons in decorators.
   - Fixed #1315: Extra newline before comment with -n + --fss.
   - Fixed #1192: `-k` or `--keep-direct-and-as-imports` option has been deprecated as it is now always on.